fix InsertOrAssign

+++ include/RED4ext/DynArray.hpp
@@ -52,7 +52,7 @@ struct DynArray
     template<class... TArgs>
     void Emplace(T* aPosition, TArgs&&... aArgs)
     {
-        uint32_t posIdx = aPosition - begin();
+        uint32_t posIdx = static_cast<uint32_t>(aPosition - begin());
         uint32_t newSize = size + 1;
         if (newSize > capacity)
         {
